requirements.txt
# Django Core
Django==5.1.1
djangorestframework==3.15.2

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.4.0

# Background Tasks
celery==5.4.0
redis==5.0.8

# File Storage
django-storages==1.14.4
boto3==1.35.16

# Payment Processing
stripe==10.8.0

# Push Notifications (Firebase Cloud Messaging)
firebase-admin>=6.2.0

# Real-time Communication
channels==4.1.0
channels-redis==4.2.0

# Admin Panel
django-unfold>=0.31.0

# Filtering
django-filter==24.3

# Phone Number Validation
django-phonenumber-field==8.0.0
phonenumbers==8.13.47

# Development Tools
django-extensions==3.2.3
python-decouple==3.8

# Testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1

# API Documentation
drf-spectacular==0.27.2

# Utilities
Pillow==10.4.0
requests==2.32.3